feat : 식단기록 화면 작업
